+++ addOns/zest/src/main/java/org/zaproxy/zap/extension/zest/menu/ZestAddAssignPopupMenu.java
@@ -24,6 +24,7 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import org.mozilla.zest.core.v1.ZestAssignCalc;
 import org.mozilla.zest.core.v1.ZestAssignFieldValue;
+import org.mozilla.zest.core.v1.ZestAssignFromElement;
 import org.mozilla.zest.core.v1.ZestAssignRandomInteger;
 import org.mozilla.zest.core.v1.ZestAssignRegexDelimiters;
 import org.mozilla.zest.core.v1.ZestAssignReplace;
@@ -130,6 +131,7 @@ public class ZestAddAssignPopupMenu extends ExtensionPopupMenuItem {
         createPopupAddActionMenu(
                 parent, child, req, new ZestAssignReplace(null, text, null, false, false));
         createPopupAddActionMenu(parent, child, req, new ZestAssignCalc());
+        createPopupAddActionMenu(parent, child, req, new ZestAssignFromElement());
     }
 
     private void createPopupAddActionMenu(